标签相关的控制类的修改，user表的修改（经纬度的字段类型）等。

+++ src/main/java/com/neo/bookhouse/entity/User.java
@@ -32,8 +32,8 @@ public class User {
 
     private Integer userKind;
     
-    private Integer userLongitude;   //经度
+    private Double userLongitude;   //经度
     
-    private Integer userLatitude;   //纬度
+    private Double userLatitude;   //纬度
 }
